# Use scan instead of ReadAt on index rebuild