Surprisingly (?) 30% faster with the same memory consumption
```
benchmark                   old ns/op      new ns/op      delta
BenchmarkRebuildIndex-4     2551647890     1725623209     -32.37%

benchmark                   old allocs     new allocs     delta
BenchmarkRebuildIndex-4     245557         245540         -0.01%

benchmark                   old bytes     new bytes     delta
BenchmarkRebuildIndex-4     48630896      48642880      +0.02%
```